refactor: extract draft file upload swagger patch into monkeyPatchDraftFileUploadSwagger
Moves the hand-written multipart/form-data requestBody override out of
createSwaggerDoc.ts into its own clearly-named file, per review feedback.
Also replaces the silent fallback with a thrown error so the patch fails
loudly if the writeDraftFile contract route is ever renamed.
